legal-demystifier/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.66.0
vertexai==0.6.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
